refactor: move commands into their own package
